ADD: Improve deletion mechanism to keep prerelease files.

+++ semversioner/storage.py
@@ -110,10 +110,22 @@ class SemversionerFileSystemStorage(SemversionerStorage):
     def remove_all_changesets(self) -> None:
         click.echo("Removing '" + self.next_release_path + "' directory.")
 
+        delete_dir = True
+
         for filename in os.listdir(self.next_release_path):
             full_path = os.path.join(self.next_release_path, filename)
-            os.remove(full_path)
-        os.rmdir(self.next_release_path)
+            changeset = None
+            with open(full_path) as f:
+                dict = json.load(f)
+                changeset = Changeset(**dict)
+
+            if changeset.pre is None:
+                os.remove(full_path)
+            else:
+                delete_dir = False
+
+        if delete_dir:
+            os.rmdir(self.next_release_path)
 
     def list_changesets(self) -> List[Changeset]:
         changes: List[Changeset] = []
